move state from Cart component to App and add select dropdown to sort by price or title
