+++ src/App.css
@@ -50,11 +50,19 @@
 
 
 .products-grid {
+  position: relative;
   width: 90%;
   margin: auto;
   min-height: 600px;
-  padding-top: 100px;
+  padding-top: 70px;
   color: white;
+}
+
+.sort-dropdown {
+  position: absolute;
+  top: 10px;
+  left: 90%;
+  transform: translateX(-50%);
 }
 
 .product button {
@@ -143,6 +151,7 @@ nav h3 {
   max-width: 1000px;
   width: 90%;
   margin: auto;
+  margin-top: 70px;
   background-color: white;
   padding: 20px;
 }
